requirements.txt
django
django-filter
djangorestframework
psycopg2